// client-monorepo/libs/applets/insurance/src/lib/app/features/equipment/api/models/renewal/order.model.ts
import { RegisterTypes } from '../../../routes/used/steps/used-device-info/models/used-register-types.model';
import { UsedPricingTypeModel } from '../../../routes/used/steps/used-pricing/models/used-pricing-type-model';
import { TabChangeModel } from '../../../routes/used/steps/used-upload-file/model/tab-change.model';

export interface OrderModel {
  orderId?: string;
  deviceId?: string;
  oldPolicyNumber?: number;
  policyNumber?: string;
  providerId?: string;
  firstName?: string;
  lastName?: string;
  mobile?: string;
  displayPrice?: string;
  isUpgEnabled?: boolean;
  productId?: string;
  productName?: string;
  productCategory?: string;
  productBrand?: string;
  productBrandId?: string;
  productModel?: string;
  announcedPrice?: number;
  fixedWageAmount?: number;
  serialNumber?: string;
  discountCode?: string;
  discountAmount?: number;
  discountValue?: number;
  voucherId?: string;
  discountDescription?: string;
  wagePercent?: number;
  wageAmount?: number;
  taxAmount?: number;
  totalAmount?: number;
  payableAmount?: number;
  orderDeadLineDays?: number;
  currentState?: number;
  isPayable?: boolean;
  paymentResultUrl: string;
  durationUnit?: string;
  durationValue?: number;
  isValidSerialNumber?: boolean;
  isValidDiscountCode?: boolean;
  unUsableReasonSerialNumber?: string;
  unUsableReasonDiscountCode?: string;
  hasSuccessPayment?: boolean;
  newPolicyIssuedAt: number;
  newPolicyExpiredAt: number;
  payTrackingCode?: string;
  gatewayName?: string;
  templateId?: string;
  templateTitle?: string;
  coverages?: string[];
  pdfUrl?: string;
  policyExpiredAt?: Date;
  nationalCode: string;
  lastTimeViewedAt?: Date;
  paymentRequestedByDesktop: boolean;
  uniqueCode: string;
  productModelId?: string;
  registerType?: RegisterTypes;
  paymentTicketTypeTitle?: 'WALLET' | 'BPG' | 'IPG';
  pricingType?: UsedPricingTypeModel;
  deviceDocuments?: { url: string, type: number, fileName: string }[];
  phoneType?: TabChangeModel;
}